+++ R/create_expectations.R
@@ -712,12 +712,12 @@ create_expectations_side_effect <- function(side_effects, name = NULL,
   assert_collection <- checkmate::makeAssertCollection()
   checkmate::assert_list(
     x = side_effects, all.missing = FALSE,
-    len = 4, add = assert_collection
+    len = 5, add = assert_collection
   )
   checkmate::assert_names(
     x = names(side_effects),
     identical.to = c(
-      "error", "warnings",
+      "error", "error_class", "warnings",
       "messages", "has_side_effects"
     ),
     type = "named"
@@ -738,28 +738,45 @@ create_expectations_side_effect <- function(side_effects, name = NULL,
     name <- deparse(substitute(side_effects))
   }
 
-  expectations <- list()
+  call_name <- name
+  name <- create_output_var_name("side_effects_", test_id)
+
+  # Create assignment string
+  assign_string <- create_assignment_strings(
+    call_name = paste0("xpectr::capture_side_effects(", call_name, ", reset_seed = TRUE)"),
+    new_name = name,
+    evaluate_once = TRUE,
+    comment = "# Assigning side effects")
+
+  expectations <- list(assign_string)
 
   if (!is.null(side_effects$error)) {
-    expectations <- c(expectations, list(
-      create_expect_side_effect(
-        name, side_effects$error,
-        side_effect_type = "error",
-        spaces = 2,
-        strip = strip
-      )
-    ))
+
+    err_expectation <- create_equality_expectation(
+      data = side_effects, name = name,
+      prefix = "",
+      suffix = "[['error']]",
+      add_strip = strip,
+      add_fixed = TRUE,
+      indentation = indentation
+    )
+
+    err_class_expectation <- create_equality_expectation(
+      data = side_effects, name = name,
+      prefix = "",
+      suffix = "[['error_class']]",
+      add_strip = strip,
+      add_fixed = TRUE,
+      indentation = indentation
+    )
+
+    expectations <- c(
+      expectations,
+      err_expectation,
+      err_class_expectation
+    )
+
   } else {
-
-    call_name <- name
-    name <- create_output_var_name("side_effects_", test_id)
-
-    # Create assignment string
-    assign_string <- create_assignment_strings(
-      call_name = paste0("xpectr::capture_side_effects(", call_name, ", reset_seed = TRUE)"),
-      new_name = name,
-      evaluate_once = TRUE,
-      comment = "# Assigning side effects")
 
     msg_expectation <- create_equality_expectation(
       data = side_effects, name = name,
@@ -779,7 +796,6 @@ create_expectations_side_effect <- function(side_effects, name = NULL,
 
     expectations <- c(
       expectations,
-      assign_string,
       warns_expectation,
       msg_expectation)
   }
@@ -789,7 +805,8 @@ create_expectations_side_effect <- function(side_effects, name = NULL,
     create_test_comment("side effects", indentation = indentation,
                         create_comment = add_comments),
     expectations
-  )
+  ) %>% unlist() %>%
+    as.list()
 }