+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Extracts error message with `conditionMessage()` instead of `$message`.
 
+* Adds `indentation = 2` to generated code from `initializeGXSFunctionAddin()`.
+
 # xpectr 0.2.0
 
 * Adds `initializeGXSFunctionAddin()` addin for initializing a `gxs_function()` call for a selected function.