+++ NEWS.md
@@ -1,4 +1,4 @@
-# xpectr 0.2.0.9000
+# xpectr 0.3.0
 
 * Breaking: `element_classes()` only returns the first class string per element.